% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSWPEGcolors.R
\name{SSWPEGcolors}
\alias{SSWPEGcolors}
\title{SSWPEG color palette}
\usage{
SSWPEGcolors(x = "all")
}
\arguments{
\item{x}{colors. Defaults to all.}
}
\description{
Use SSWPEG colors in graphs.
}
\examples{
hist(iris$Petal.Length, col = SSWPEGcolors())
plot(iris$Petal.Length, iris$Sepal.Length, col = SSWPEGcolors(), pch = 19)
}
\keyword{SSWPEG}
\keyword{colors}
